fix: broken init script
Turns out that the init script must not import aliased paths, otherwise
it breaks
