+++ dev.ts
@@ -1,2 +1,2 @@
-import { dev } from "./src/dev/mod.ts";
+import { dev } from "./src/dev/dev_command.ts";
 export default dev;
